+++ bot/commands/bot/admin.js
@@ -1,5 +1,6 @@
 'use strict';
 var settings = require(process.cwd() + '/private/settings.js');
+var historyStore = require(process.cwd()+ '/bot/store/history.js');
 var _ = require('lodash');
 
 settings.APPROVED_USERS.push(settings.OWNER);
@@ -52,7 +53,10 @@ module.exports = function(bot, db, data) {
         bot.myconfig.muted = true;
         bot.sendChat = function(x){return;};
       }
-      
+      break;
+    case 'history':
+      bot.sendChat('ok, updating my room playlist history');
+      historyStore.init(bot);
       break;
     case 'unmute':
       if (bot.myconfig.muted) {